dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "db-dtypes>=1.5.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.36.2",
    "pandas>=2.3.3",
    "polars>=1.38.1",
    "streamlit>=1.55.0",
    "vegafusion>=2.0.3",
]
